Resolved comments:     -moved all services to services folder     -renamed LoadaingPage to LoadingSpinner
