--- Web/src/pages/Auth/AuthDataService.jsx
@@ -1,9 +1,0 @@
-import { callApi } from '../../DataService';
-
-export async function registerUser(data) {
-    return callApi({ method: 'POST', route: 'auth/register', data });
-}
-
-export async function loginUser(data) {
-    return callApi({ method: 'POST', route: 'auth/login', data });
-}
