+++ Web/src/services/ProfileDataService.jsx
@@ -1,0 +1,13 @@
+import { callApi } from './DataService';
+
+export async function getProfile({ route, id }) {
+    return callApi('POST', route, { idKorisnik: id });
+}
+
+export async function follow({ idUser, idFollow }) {
+    return callApi('POST', 'follow', { idKorisnik: idUser, idZapratiti: idFollow });
+}
+
+export async function unfollow({ idUser, idFollow }) {
+    return callApi('DELETE', 'follow', { idKorisnik: idUser, idZapratiti: idFollow });
+}
